fix more -Werror=sign-compare errors in device code

+++ cpp/include/cudf/strings/detail/strings_column_factories.cuh
@@ -52,7 +52,7 @@ std::unique_ptr<column> make_strings_column(IndexPairIterator begin,
   };
   size_t bytes = thrust::transform_reduce(
     rmm::exec_policy(stream), begin, end, size_checker, 0, thrust::plus<size_t>());
-  CUDF_EXPECTS(bytes < std::numeric_limits<size_type>::max(),
+  CUDF_EXPECTS(bytes < static_cast<std::size_t>(std::numeric_limits<size_type>::max()),
                "total size of strings is too large for cudf column");
 
   // build offsets column from the strings sizes
